Added .dockerignore
